+++ src/dxvk/rtx_render/graph/components/texture_hash_checker.h
@@ -35,7 +35,7 @@ namespace components {
 
 #define LIST_OUTPUTS(X) \
   X(RtComponentPropertyType::Bool, false, isUsed, "Is Used", "True if the texture hash was used in the current frame.") \
-  X(RtComponentPropertyType::Uint32, 0, usageCount, "Usage Count", "Number of times the texture hash was used in the current frame.")
+  X(RtComponentPropertyType::Float, 0.0f, usageCount, "Usage Count", "Number of times the texture hash was used in the current frame.")
 
 REMIX_COMPONENT( \
   /* the Component name */ TextureHashChecker, \
@@ -62,7 +62,7 @@ void TextureHashChecker::updateRange(const Rc<DxvkContext>& context, const size_
     bool isUsed = count > 0;
     
     m_isUsed[i] = isUsed;
-    m_usageCount[i] = count;
+    m_usageCount[i] = float(count);
   }
 }
